net: mqtt: discard a hack for original paho code
The setsockopt() less than 1 dsec timeout issue is fixed by
this commit from NuttX:

commit 314ec87c86f6d1bc2ad5938ed020d1846d966e2a
Author: Gregory Nutt <gnutt@nuttx.org>
Date:   Fri May 29 07:17:15 2015 -0600

Networking: setsockopt() fails when setting timeouts to values less than 100 msec.
That is because the timeout is limited to stops of 1 decisecond and
because the conversion of structure timeval was truncating the microsecond
remainder. The utility net_timeval2dsec now accespts and option to determin
how it handles the remainder: truncate, discarding the remainder,
use the remainder to round to the closed decisecond value, or use any non-zero
remainder to the next larger whole decisecond value.

For the setsockopt() case, it will always use the final option.
This means that a time of zero (i.e., not timeout) can only
occur if the value zero was explicated provided to setsockopt().